The diff shows the changes made to the SaveEntryAction file.
In the first change, the code checks if the thisWeekEntries variable is set and if it is equal to 6. If it is, the code sets the is_rest_day variable to false and the is_validated variable to true.

In the second change, the code checks if the thisWeekEntries variable is set and if it is equal to 6. If it is, the code sets the is_rest_day variable to false and the is_validated variable to true.

In the third change, the code checks if the thisWeekEntries variable is set and if it is equal to 6. If it is, the code sets the is_rest_day variable to false and the is_validated variable to true.